"Update chart dataset and options in ChartDataManager"
This commit updates the chart dataset and options in the ChartDataManager class. The 'views' dataset has been modified to include additional properties such as backgroundColor, pointBackgroundColor, etc. The 'scales' and 'plugins' sections have been added to the 'options' object. The 'animation' property has also been added to the 'options' object.